# bff/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/semester/getSemester": {
            "post": {
                "description": "获取当前日期所属学期",
                "tags": [
                    "semester"
                ],
                "summary": "获取当前日期所属学期",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "获取学期信息请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/content.GetSemesterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/content.GetSemesterResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/semester/saveSemester": {
            "post": {
                "description": "保存学期信息",
                "tags": [
                    "semester"
                ],
                "summary": "保存学期信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "保存学期信息请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/content.SaveSemesterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/web.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "content.GetSemesterRequest": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                }
            }
        },
        "content.GetSemesterResponse": {
            "type": "object",
            "properties": {
                "semester": {
                    "type": "string"
                }
            }
        },
...
        "content.SaveSemesterRequest": {
            "type": "object",
            "required": [
                "end_date",
                "semester",
                "start_date"
            ],
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "semester": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
...
        "web.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}